requirements.txt
streamlit>=1.36
tensorflow>=2.12
opencv-python
numpy
pillow
matplotlib
scikit-learn